Moved Noise.cs to its own folder
